relocate files from 'drivers/board/'
