+++ Marlin/lib/h32_core/main/main.cpp
@@ -1,5 +1,5 @@
 #include "../Arduino.h"
-#include "../drivers/board/init.h"
+#include "init.h"
 
 int main(void)
 {
